requirements.txt
asgiref==3.3.4
astroid==2.5.3
colorama==0.4.4
Django==3.2
django-adminlte-3==0.1.6
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-mptt==0.12.0
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mysqlclient==2.0.3
Pillow==8.2.0
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
